api airport

+++ service/src/main/java/com/eligasht/service/model/flight/response/searchFlight/TotalFare.java
@@ -8,16 +8,16 @@ public class TotalFare {
 
     @SerializedName("Amount")
     @Expose
-    private Integer amount;
+    private double amount;
     @SerializedName("CurrencyCode")
     @Expose
     private String currencyCode;
 
-    public Integer getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
